Duplicate device hack
Preface:  I've never written a line of Node before.  

All I'm doing here is keeping track of what devices have been added so far, and skipping anything that's already been found.  "First one wins".  This is a complete and total hack, and probably should not be merged.

The proper fix is somewhere in node-sonos in the LogicalDevice.search method, but that was way over my head.  With any luck, this might fix issue #114 - that's the itch I'm trying to scratch.  There's a large chance it may not.